fix unsteady output, results directories
