+++ funtofem/interface/utils/__init__.py
@@ -19,5 +19,7 @@ fun3d_loader = importlib.util.find_spec("fun3d")
 if fun3d_loader is not None:
     from .fun3d_grid_interface import *
 
+from .test_utils import *
+
 # need to be updated and therefore commmented out for now
 # if fun3d_loader is not None: from .fun3d_client import *
