+++ tests/unit_tests/framework_unsteady/test_tacs_interface_unsteady.py
@@ -12,7 +12,7 @@ from funtofem.interface import (
     TestResult,
 )
 from funtofem.driver import FUNtoFEMnlbgs, TransferSettings
-from bdf_test_utils import elasticity_callback, thermoelasticity_callback
+from _bdf_test_utils import elasticity_callback, thermoelasticity_callback
 
 np.random.seed(123456)
