+++ tests/unit_tests/shape/test_tacs_driver_unsteady_coordinate.py
@@ -12,9 +12,9 @@ from funtofem.interface import (
     TacsIntegrationSettings,
     CoordinateDerivativeTester,
 )
-from funtofem.driver import TransferSettings, FUNtoFEMnlbgs
+from funtofem.driver import TacsOnewayDriver, TransferSettings, FUNtoFEMnlbgs
 
-from bdf_test_utils import elasticity_callback, thermoelasticity_callback
+from _bdf_test_utils import elasticity_callback, thermoelasticity_callback
 import unittest
 
 np.random.seed(123456)
@@ -34,16 +34,15 @@ if comm.rank == 0:  # make the results folder if doesn't exist
     if not os.path.exists(results_folder):
         os.mkdir(results_folder)
 
-in_github_workflow = bool(os.getenv("GITHUB_ACTIONS"))
 
-
-@unittest.skipIf(in_github_workflow, "still in development")
-class TestFuntofemDriverUnsteadyAeroCoordinate(unittest.TestCase):
-    FILENAME = "f2f-unsteady-aero-coord.txt"
+@unittest.skipIf(
+    not complex_mode, "only testing coordinate derivatives with complex step"
+)
+class TestTacsDriverUnsteadyCoordinate(unittest.TestCase):
+    FILENAME = "testaero-tacsdriver-unsteady.txt"
     FILEPATH = os.path.join(results_folder, FILENAME)
 
-    # @unittest.skip("under development")
-    def test_unsteady_aero_aeroelastic(self):
+    def test_unsteady_aeroelastic(self):
         # build the model and driver
         model = FUNtoFEMmodel("wedge")
         plate = Body.aeroelastic("plate", boundary=1)
@@ -51,11 +50,9 @@ class TestFuntofemDriverUnsteadyAeroCoordinate(unittest.TestCase):
 
         # build the scenario
         scenario = Scenario.unsteady("test", steps=10).include(Function.ksfailure())
-        scenario.include(Function.lift()).include(Function.drag())
         integration_settings = TacsIntegrationSettings(
             dt=0.001, num_steps=scenario.steps
         )
-        Variable.shape("rotation").register_to(scenario)
         scenario.include(integration_settings).register_to(model)
 
         # build the tacs interface, coupled driver, and oneway driver
@@ -70,21 +67,20 @@ class TestFuntofemDriverUnsteadyAeroCoordinate(unittest.TestCase):
             callback=elasticity_callback,
             output_dir=output_folder,
         )
-        transfer_settings = TransferSettings(elastic_scheme="meld", npts=5)
-        # transfer_settings = TransferSettings(elastic_scheme="rbf", npts=5, options={
-        #     "basis function" : "multiquadric"
-        # })
+        solvers.structural._coord_deriv_override = True
+        transfer_settings = TransferSettings(npts=5)
         coupled_driver = FUNtoFEMnlbgs(
             solvers, transfer_settings=transfer_settings, model=model
         )
+        oneway_driver = TacsOnewayDriver.prime_loads(coupled_driver)
 
         epsilon = 1e-30 if complex_mode else 1e-4
-        rtol = 1e-7 if complex_mode else 1e-5
+        rtol = 1e-9 if complex_mode else 1e-5
 
         """complex step test over coordinate derivatives"""
-        tester = CoordinateDerivativeTester(coupled_driver)
-        rel_error = tester.test_aero_coordinates(
-            "funtofem_driver aero coordinate derivatives unsteady-aeroelastic",
+        tester = CoordinateDerivativeTester(oneway_driver)
+        rel_error = tester.test_struct_coordinates(
+            "tacs_driver coordinate derivatives unsteady-aeroelastic",
             status_file=self.FILEPATH,
             epsilon=epsilon,
             complex_mode=complex_mode,
@@ -92,8 +88,7 @@ class TestFuntofemDriverUnsteadyAeroCoordinate(unittest.TestCase):
         assert abs(rel_error) < rtol
         return
 
-    # @unittest.skip("under development")
-    def test_unsteady_aero_aerothermal(self):
+    def test_unsteady_aerothermal(self):
         # build the model and driver
         model = FUNtoFEMmodel("wedge")
         plate = Body.aerothermal("plate", boundary=1)
@@ -101,11 +96,9 @@ class TestFuntofemDriverUnsteadyAeroCoordinate(unittest.TestCase):
 
         # build the scenario
         scenario = Scenario.unsteady("test", steps=10).include(Function.temperature())
-        scenario.include(Function.lift()).include(Function.drag())
         integration_settings = TacsIntegrationSettings(
             dt=0.001, num_steps=scenario.steps
         )
-        Variable.shape("rotation").register_to(scenario)
         scenario.include(integration_settings).register_to(model)
 
         # build the tacs interface, coupled driver, and oneway driver
@@ -120,18 +113,20 @@ class TestFuntofemDriverUnsteadyAeroCoordinate(unittest.TestCase):
             callback=thermoelasticity_callback,
             output_dir=output_folder,
         )
+        solvers.structural._coord_deriv_override = True
         transfer_settings = TransferSettings(npts=5)
         coupled_driver = FUNtoFEMnlbgs(
             solvers, transfer_settings=transfer_settings, model=model
         )
+        oneway_driver = TacsOnewayDriver.prime_loads(coupled_driver)
 
         epsilon = 1e-30 if complex_mode else 1e-4
-        rtol = 1e-7 if complex_mode else 1e-5
+        rtol = 1e-9 if complex_mode else 1e-5
 
         """complex step test over coordinate derivatives"""
-        tester = CoordinateDerivativeTester(coupled_driver)
-        rel_error = tester.test_aero_coordinates(
-            "funtofem_driver aero coordinate derivatives unsteady-aerothermal",
+        tester = CoordinateDerivativeTester(oneway_driver)
+        rel_error = tester.test_struct_coordinates(
+            "tacs_driver coordinate derivatives unsteady-aerothermal",
             status_file=self.FILEPATH,
             epsilon=epsilon,
             complex_mode=complex_mode,
@@ -139,8 +134,7 @@ class TestFuntofemDriverUnsteadyAeroCoordinate(unittest.TestCase):
         assert abs(rel_error) < rtol
         return
 
-    # @unittest.skip("under development")
-    def test_unsteady_aero_aerothermoelastic(self):
+    def test_unsteady_aerothermoelastic(self):
         # build the model and driver
         model = FUNtoFEMmodel("wedge")
         plate = Body.aerothermoelastic("plate", boundary=1)
@@ -148,12 +142,10 @@ class TestFuntofemDriverUnsteadyAeroCoordinate(unittest.TestCase):
 
         # build the scenario
         scenario = Scenario.unsteady("test", steps=10).include(Function.temperature())
-        scenario.include(Function.ksfailure(ks_weight=5.0))
-        scenario.include(Function.lift()).include(Function.drag())
+        scenario.include(Function.temperature()).include(Function.lift())
         integration_settings = TacsIntegrationSettings(
             dt=0.001, num_steps=scenario.steps
         )
-        Variable.shape("rotation").register_to(scenario)
         scenario.include(integration_settings).register_to(model)
 
         # build the tacs interface, coupled driver, and oneway driver
@@ -168,18 +160,20 @@ class TestFuntofemDriverUnsteadyAeroCoordinate(unittest.TestCase):
             callback=thermoelasticity_callback,
             output_dir=output_folder,
         )
+        solvers.structural._coord_deriv_override = True
         transfer_settings = TransferSettings(npts=5)
         coupled_driver = FUNtoFEMnlbgs(
             solvers, transfer_settings=transfer_settings, model=model
         )
+        oneway_driver = TacsOnewayDriver.prime_loads(coupled_driver)
 
         epsilon = 1e-30 if complex_mode else 1e-4
-        rtol = 1e-7 if complex_mode else 1e-5
+        rtol = 1e-9 if complex_mode else 1e-5
 
         """complex step test over coordinate derivatives"""
-        tester = CoordinateDerivativeTester(coupled_driver)
-        rel_error = tester.test_aero_coordinates(
-            "funtofem_driver aero coordinate derivatives unsteady-aerothermoelastic",
+        tester = CoordinateDerivativeTester(oneway_driver)
+        rel_error = tester.test_struct_coordinates(
+            "tacs_driver coordinate derivatives unsteady-aerothermoelastic",
             status_file=self.FILEPATH,
             epsilon=epsilon,
             complex_mode=complex_mode,
@@ -187,11 +181,14 @@ class TestFuntofemDriverUnsteadyAeroCoordinate(unittest.TestCase):
         assert abs(rel_error) < rtol
         return
 
-    @unittest.skip("need to fix multi-scenario coord derivatives")
-    def test_unsteady_aero_multiscenario_aerothermoelastic(self):
+    @unittest.skip("have to fix multiscenario case")
+    def test_unsteady_multiscenario_aerothermoelastic(self):
         # build the model and driver
         model = FUNtoFEMmodel("wedge")
         plate = Body.aerothermoelastic("plate", boundary=1)
+        Variable.structural("thickness").set_bounds(
+            lower=0.01, value=0.1, upper=1.0
+        ).register_to(plate)
         plate.register_to(model)
 
         # build the scenario
@@ -220,24 +217,26 @@ class TestFuntofemDriverUnsteadyAeroCoordinate(unittest.TestCase):
             callback=thermoelasticity_callback,
             output_dir=output_folder,
         )
-        solvers.structural.can_skip_coordinates = False
         transfer_settings = TransferSettings(npts=5)
         coupled_driver = FUNtoFEMnlbgs(
             solvers, transfer_settings=transfer_settings, model=model
         )
+        oneway_driver = TacsOnewayDriver.prime_loads(coupled_driver)
 
-        rtol = 1e-9
+        epsilon = 1e-30 if complex_mode else 1e-4
+        rtol = 1e-9 if complex_mode else 1e-5
 
         """complex step test over coordinate derivatives"""
-        tester = CoordinateDerivativeTester(coupled_driver)
-        rel_error = tester.test_aero_coordinates(
-            "funtofem_driver aero coordinate derivatives multiscenario, unsteady-aerothermoelastic"
+        tester = CoordinateDerivativeTester(oneway_driver)
+        rel_error = tester.test_struct_coordinates(
+            "tacs_driver coordinate derivatives multiscenario, unsteady-aerothermoelastic",
+            status_file=self.FILEPATH,
+            epsilon=epsilon,
+            complex_mode=complex_mode,
         )
         assert abs(rel_error) < rtol
         return
 
 
 if __name__ == "__main__":
-    open(TestFuntofemDriverUnsteadyAeroCoordinate.FILEPATH, "w").close()
-    complex_mode = True
     unittest.main()
